requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
evaluate>=0.4.1
sentence-transformers>=2.7.0
peft>=0.10.0

# Classical ML + metrics
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0

# Code analysis
networkx>=3.2
faiss-cpu>=1.7.4

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Data + IO
requests>=2.31.0
tqdm>=4.66.0
pyarrow>=15.0.0
python-dotenv>=1.0.0

# Notebook
jupyter>=1.0.0
ipywidgets>=8.1.0
